apps/eleme-auto-activity-signup/requirements.txt
: | # Python依赖
playwright>=1.40.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
python-dotenv>=1.0.0
